backend/requirements.txt
flask
flask-cors
flask_jwt_extended
python-dotenv
pymongo